pyproject.toml
[tool.poetry]
name = "ftp2webdav"
version = "0.1"
description = "An FTP server that forwards uploaded files to a WebDAV server"
authors = ["Tomas Bayer <mail@tomas.berlin>"]
readme = "README.md"
license = "MIT"

[tool.poetry.scripts]
ftp2webdav = 'ftp2webdav.cli:run'

[tool.poetry.dependencies]
python = "^3.9"
cerberus = "^1.3.4"
easywebdav = "^1.2.0"
ftprelay = "^0.1"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
